[refactor] User 안에 있던 MailSender -> Notification으로 기능 분리
