+++ src/main/java/kr/co/pawong/pwbe/notification/application/port/in/CustomMailSenderUseCase.java
@@ -1,0 +1,7 @@
+package kr.co.pawong.pwbe.notification.application.port.in;
+
+public interface CustomMailSenderUseCase {
+
+    void sendEmail(String toEmail, String title, String text);
+
+}
