+++ src/main/java/kr/co/pawong/pwbe/notification/application/port/in/MailUseCase.java
@@ -1,0 +1,9 @@
+package kr.co.pawong.pwbe.notification.application.port.in;
+
+public interface MailUseCase {
+
+    void sendCodeToEmail(String toEmail);
+
+    String verifiedCode(String email, String authCode);
+
+}
