+++ src/main/java/kr/co/pawong/pwbe/user/adapter/in/api/EmailController.java
@@ -1,7 +1,6 @@
 package kr.co.pawong.pwbe.user.adapter.in.api;
 
-import jakarta.annotation.security.PermitAll;
-import kr.co.pawong.pwbe.user.application.port.in.AuthUseCase;
+import kr.co.pawong.pwbe.notification.application.port.in.MailUseCase;
 import kr.co.pawong.pwbe.user.application.port.in.QueryUserDataUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -18,25 +17,18 @@ import org.springframework.web.bind.annotation.RestController;
 public class EmailController {
 
     private final QueryUserDataUseCase queryUserDataUseCase;
-    private final AuthUseCase authUseCase;
-
-    @PermitAll
-    @GetMapping("/is-exist")
-    public ResponseEntity<Boolean> isEmailExist(@RequestParam(value = "email", required = false) String email) {
-        return ResponseEntity.ok(queryUserDataUseCase.isEmailExist(email));
-    }
+    private final MailUseCase mailUseCase;
 
     @PostMapping("/six")
     public ResponseEntity sendMessage(@RequestParam("email") String email) {
-        authUseCase.sendCodeToEmail(email);
-
+        mailUseCase.sendCodeToEmail(email);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/verifications")
     public ResponseEntity verificationEmail(@RequestParam("email") String email,
             @RequestParam("code") String authCode) {
-        String response = authUseCase.verifiedCode(email, authCode);
+        String response = mailUseCase.verifiedCode(email, authCode);
 
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
