--- src/main/java/kr/co/pawong/pwbe/user/application/port/in/EmailUseCase.java
@@ -1,9 +1,0 @@
-package kr.co.pawong.pwbe.user.application.port.in;
-
-import org.springframework.mail.SimpleMailMessage;
-
-public interface EmailUseCase {
-
-    void sendEmail(String toEmail, String title, String text);
-
-}
